Add test-support feature to reduce boilerplate for common test and example code
