+++ Cargo.toml
@@ -37,13 +37,14 @@ serde-this-or-that = "0.4.2"
 time = { version = "0.3.36", features = ["serde", "macros", "parsing"] }
 rust_decimal = { version = "1.35.0", features = ["serde-with-str", "serde-with-float"] }
 rust_decimal_macros = "1.35.0"
+tracing-subscriber = { version = "0.3.18", optional = true }
 
 [dev-dependencies]
 wiremock = "0.6.1"
-tracing-subscriber = "0.3.18"
 ws-mock = "0.2.0"
 
 [features]
-default = []
+default = ["test-support"]
 debug-inbound = []
 debug-outbound = []
+test-support = ["dep:tracing-subscriber"]
