+++ examples/live_open_orders_request.rs
@@ -4,12 +4,14 @@ use kraken_async_rs::clients::kraken_client::KrakenClient;
 use kraken_async_rs::crypto::nonce_provider::{IncreasingNonceProvider, NonceProvider};
 use kraken_async_rs::request_types::OpenOrdersRequest;
 use kraken_async_rs::secrets::secrets_provider::{EnvSecretsProvider, SecretsProvider};
+use kraken_async_rs::test_support::set_up_logging;
 use std::sync::Arc;
 use tokio::sync::Mutex;
 
 /// This retrieves the user's open orders, provided the API key and secret are available in env vars.
 #[tokio::main]
 async fn main() {
+    set_up_logging("open_orders_request.log");
     // note that this will fail if you don't have your key and secret set to these env vars
     // eg `export KRAKEN_KEY="YOUR-API-KEY"`, ...
     let secrets_provider: Box<Arc<Mutex<dyn SecretsProvider>>> = Box::new(Arc::new(Mutex::new(
