+++ examples/live_query_orders_info.rs
@@ -3,11 +3,9 @@ use kraken_async_rs::clients::kraken_client::KrakenClient;
 use kraken_async_rs::crypto::nonce_provider::{IncreasingNonceProvider, NonceProvider};
 use kraken_async_rs::request_types::OrderRequest;
 use kraken_async_rs::secrets::secrets_provider::{EnvSecretsProvider, SecretsProvider};
-use std::fs::File;
+use kraken_async_rs::test_support::set_up_logging;
 use std::sync::Arc;
 use tokio::sync::Mutex;
-use tracing_subscriber::layer::SubscriberExt;
-use tracing_subscriber::{fmt, Registry};
 
 /// This retrieves the user's open orders, provided the API key and secret are available in env vars.
 #[tokio::main]
@@ -31,24 +29,4 @@ async fn main() {
     for (id, order) in response.unwrap().result.unwrap() {
         println!("{id}: {order:?}");
     }
-}
-
-fn set_up_logging(filename: &str) {
-    let subscriber = Registry::default()
-        .with(
-            fmt::Layer::default()
-                .with_ansi(false)
-                .with_writer(get_log_file(filename)),
-        )
-        .with(fmt::Layer::default().pretty().with_writer(std::io::stdout));
-
-    tracing::subscriber::set_global_default(subscriber).unwrap();
-}
-
-fn get_log_file(filename: &str) -> File {
-    File::options()
-        .append(true)
-        .create(true)
-        .open(filename)
-        .expect("failed to open test log file!")
 }
