+++ examples/live_wss_executions_v2.rs
@@ -5,7 +5,7 @@ use kraken_async_rs::secrets::secrets_provider::{EnvSecretsProvider, SecretsProv
 use kraken_async_rs::test_support::set_up_logging;
 use kraken_async_rs::wss::v2::base_messages::{Message, WssMessage};
 use kraken_async_rs::wss::v2::kraken_wss_client::KrakenWSSClient;
-use kraken_async_rs::wss::v2::user_data_messages::BalancesSubscription;
+use kraken_async_rs::wss::v2::user_data_messages::{BalancesSubscription, ExecutionSubscription};
 use std::fs::File;
 use std::sync::Arc;
 use std::time::Duration;
@@ -20,12 +20,11 @@ use tracing_subscriber::{fmt, Registry};
 /// the Balances channel, listening for a snapshot and any updates to balances.
 #[tokio::main]
 async fn main() {
-    set_up_logging("wss_balances_v2.log");
+    set_up_logging("wss_executions_v2.log");
 
     let secrets_provider: Box<Arc<Mutex<dyn SecretsProvider>>> = Box::new(Arc::new(Mutex::new(
         EnvSecretsProvider::new("KRAKEN_KEY", "KRAKEN_SECRET"),
     )));
-
     let nonce_provider: Box<Arc<Mutex<dyn NonceProvider>>> =
         Box::new(Arc::new(Mutex::new(IncreasingNonceProvider::new())));
     let mut kraken_client = CoreKrakenClient::new(secrets_provider, nonce_provider);
@@ -37,8 +36,10 @@ async fn main() {
     let mut client = KrakenWSSClient::new();
     let mut kraken_stream = client.connect_auth::<WssMessage>().await.unwrap();
 
-    let balances_params = BalancesSubscription::new(token);
-    let subscription = Message::new_subscription(balances_params, 0);
+    let mut exeuction_params = ExecutionSubscription::new(token);
+    exeuction_params.snapshot = Some(true);
+    exeuction_params.snapshot_trades = Some(true);
+    let subscription = Message::new_subscription(exeuction_params, 0);
 
     let result = kraken_stream.send(&subscription).await;
     assert!(result.is_ok());
