+++ examples/live_wss_trading_v2.rs
@@ -5,6 +5,7 @@ use kraken_async_rs::crypto::secrets::Token;
 use kraken_async_rs::request_types::{SelfTradePrevention, TimeInForceV2};
 use kraken_async_rs::response_types::{BuySell, OrderType};
 use kraken_async_rs::secrets::secrets_provider::{EnvSecretsProvider, SecretsProvider};
+use kraken_async_rs::test_support::set_up_logging;
 use kraken_async_rs::wss::v2::base_messages::{Message, MethodMessage, ResultResponse, WssMessage};
 use kraken_async_rs::wss::v2::kraken_wss_client::KrakenWSSClient;
 use kraken_async_rs::wss::v2::trading_messages::{
@@ -146,24 +147,4 @@ fn order_edit_from_add_order_result(
         params: order_edit,
         req_id: 0,
     }
-}
-
-fn set_up_logging(filename: &str) {
-    let subscriber = Registry::default()
-        .with(
-            fmt::Layer::default()
-                .with_ansi(false)
-                .with_writer(get_log_file(filename)),
-        )
-        .with(fmt::Layer::default().pretty().with_writer(std::io::stdout));
-
-    tracing::subscriber::set_global_default(subscriber).unwrap();
-}
-
-fn get_log_file(filename: &str) -> File {
-    File::options()
-        .append(true)
-        .create(true)
-        .open(filename)
-        .expect("failed to open test log file!")
 }
