+++ src/test_support/mod.rs
@@ -1,0 +1,3 @@
+mod logging;
+
+pub use logging::*;
